+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Add redirect for when invalid link is entered
 * Fixed searches for ID64 not working
+* Added second attempt to load avatar when viewing past player